agents/repo_cloner/pyproject.toml
[project]
name = "repo-cloner-agent"
version = "0.1.0"
description = "ADK agent for cloning Git repositories"
requires-python = ">=3.9"
dependencies = [
    "google-adk>=0.0.3",
]

[tool.hatch.build.targets.wheel]
packages = ["."]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"